requirements.txt
asgiref==3.7.2
Babel==2.14.0
boto3==1.34.36
botocore==1.34.36
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap4==2023.1
cryptography==42.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.2.10
django-allauth==0.60.1
django-countries==7.2.1
django-crispy-forms==2.1
django-embed-video==1.4.9
django-phonenumber-field==7.2.0
django-storages==1.14.2
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
oauthlib==3.2.2
packaging==23.2
phonenumberslite==8.13.29
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
stripe==8.1.0
typing_extensions==4.9.0
urllib3==1.26.18
whitenoise==6.7.0
